ctrl: nrop: dangling: move object removal in pkg
Move the functions removing danglig objects
in the newly-added `dangling` pkg.
Code movement and trivial API changes only.
