+++ internal/dangling/dangling.go
@@ -21,3 +21,93 @@
 // objects are left unnecessarily lingering. In hindsight, we should probably have set a NUMAResourcesOperator own NodeGroups, or just allow more than
 // a NUMAResourcesOperator object, but that ship as sailed and now a NUMAResourcesOperator object is 1:N to NodeGroups (and the latter are not K8S objects).
 package dangling
+
+import (
+	"context"
+
+	appsv1 "k8s.io/api/apps/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/sets"
+	"k8s.io/klog/v2"
+
+	"sigs.k8s.io/controller-runtime/pkg/client"
+
+	machineconfigv1 "github.com/openshift/machine-config-operator/pkg/apis/machineconfiguration.openshift.io/v1"
+
+	nropv1 "github.com/openshift-kni/numaresources-operator/api/numaresourcesoperator/v1"
+	nodegroupv1 "github.com/openshift-kni/numaresources-operator/api/numaresourcesoperator/v1/helper/nodegroup"
+	"github.com/openshift-kni/numaresources-operator/pkg/objectnames"
+)
+
+func DeleteUnusedDaemonSets(cli client.Client, ctx context.Context, instance *nropv1.NUMAResourcesOperator, trees []nodegroupv1.Tree) []error {
+	klog.V(3).Info("Delete Daemonsets start")
+	defer klog.V(3).Info("Delete Daemonsets end")
+	var errors []error
+	var daemonSetList appsv1.DaemonSetList
+	if err := cli.List(ctx, &daemonSetList, &client.ListOptions{Namespace: instance.Namespace}); err != nil {
+		klog.ErrorS(err, "error while getting Daemonset list")
+		return append(errors, err)
+	}
+
+	expectedDaemonSetNames := sets.NewString()
+	for _, tree := range trees {
+		for _, mcp := range tree.MachineConfigPools {
+			expectedDaemonSetNames = expectedDaemonSetNames.Insert(objectnames.GetComponentName(instance.Name, mcp.Name))
+		}
+	}
+
+	for _, ds := range daemonSetList.Items {
+		if !expectedDaemonSetNames.Has(ds.Name) {
+			if isOwnedBy(ds.GetObjectMeta(), instance) {
+				if err := cli.Delete(ctx, &ds); err != nil {
+					klog.ErrorS(err, "error while deleting daemonset", "DaemonSet", ds.Name)
+					errors = append(errors, err)
+				} else {
+					klog.V(3).InfoS("Daemonset deleted", "name", ds.Name)
+				}
+			}
+		}
+	}
+	return errors
+}
+
+func DeleteUnusedMachineConfigs(cli client.Client, ctx context.Context, instance *nropv1.NUMAResourcesOperator, trees []nodegroupv1.Tree) []error {
+	klog.V(3).Info("Delete Machineconfigs start")
+	defer klog.V(3).Info("Delete Machineconfigs end")
+	var errors []error
+	var machineConfigList machineconfigv1.MachineConfigList
+	if err := cli.List(ctx, &machineConfigList); err != nil {
+		klog.ErrorS(err, "error while getting MachineConfig list")
+		return append(errors, err)
+	}
+
+	expectedMachineConfigNames := sets.NewString()
+	for _, tree := range trees {
+		for _, mcp := range tree.MachineConfigPools {
+			expectedMachineConfigNames = expectedMachineConfigNames.Insert(objectnames.GetMachineConfigName(instance.Name, mcp.Name))
+		}
+	}
+
+	for _, mc := range machineConfigList.Items {
+		if !expectedMachineConfigNames.Has(mc.Name) {
+			if isOwnedBy(mc.GetObjectMeta(), instance) {
+				if err := cli.Delete(ctx, &mc); err != nil {
+					klog.ErrorS(err, "error while deleting machineconfig", "MachineConfig", mc.Name)
+					errors = append(errors, err)
+				} else {
+					klog.V(3).InfoS("Machineconfig deleted", "name", mc.Name)
+				}
+			}
+		}
+	}
+	return errors
+}
+
+func isOwnedBy(element metav1.Object, owner metav1.Object) bool {
+	for _, ref := range element.GetOwnerReferences() {
+		if ref.UID == owner.GetUID() {
+			return true
+		}
+	}
+	return false
+}
